2014-01-30 10:34 UTC+0100 Przemyslaw Czerpak (druzus/at/poczta.onet.pl)   * contrib/hbtip/utils.c   * src/rtl/gtwvt/gtwvt.c     * casting to pacify some warnings
  * src/compiler/harbour.y
  * src/compiler/harbour.yyc
  * src/compiler/harbour.yyh
    * changed type of valChar.length from int to HB_SIZE

  * include/hbpp.h
  * src/pp/ppcore.c
    * changed type of last hb_pp_tokenBlockString() parameter
      from int * to HB_SIZE *

  * src/compiler/complex.c
    * removed unnecessary casting

+++ src/rtl/gtwvt/gtwvt.c
@@ -2524,7 +2524,7 @@ static HB_BOOL hb_gt_wvt_KeyEvent( PHB_GTWVT pWVT, UINT message, WPARAM wParam, 
                if( iFlags & HB_KF_CTRL )
                {
                   pWVT->IgnoreWM_SYSCHAR = HB_TRUE;
-                  iKey = wParam - VK_NUMPAD0 + '0';
+                  iKey = ( int ) wParam - VK_NUMPAD0 + '0';
                }
                else if( iFlags == HB_KF_ALT )
                   iFlags = 0; /* for ALT + <ASCII_VALUE_FROM_KEYPAD> */
